Finsihed Node Redesign, Added Healthbars, Bounty Animation,  Made other minor adjustments
Healthbars and a bounty animaiton have been added to enemy views,
uncomplete due to not rotating freely from enemy visuals.
Nodes have been changed to an imported material that makes use of
emission for a glow effect, pylon script ahs been altered to change this
emission colour to reflect whether a node is activated or not. Also made
change to base walls.